% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunspell.R
\name{hunspell_check_en}
\alias{hunspell_check_en}
\alias{hunspell_suggest_en}
\title{Hunspell}
\usage{
hunspell_check_en(words, ignore = character())

hunspell_suggest_en(words)
}
\arguments{
\item{words}{character vector of words to spellcheck}

\item{ignore}{additional approved words in dictionary}
}
\description{
Run hunspell spell checker. Additional dictrionaries can be downloaded from
any OpenOffice \href{mirror}{http://ftp.snt.utwente.nl/pub/software/openoffice/contrib/dictionaries/}.
}

